refactor(research_tracker/idea_extraction): promote to directory, split tests (#1266)
idea_extraction.rs (431 LOC) is promoted to a directory module:
- src/research_tracker/idea_extraction/mod.rs: 222 LOC (down from 431)
- src/research_tracker/idea_extraction/tests.rs: new (~209 LOC)

Pure refactor: zero logic changes, all 16 tests pass.

Closes 5/70 of #1266 (this loop iteration), 13 PRs cumulative since #1285.
